# swagger/all.yaml
openapi: 3.0.0
info:
  description: |
    A more complex version of the race cars API.
  version: "1.0.0"
  title: Race Cars Management
  contact:
    email: loi.tra@9thwonder.com
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/loi.tra/my-first-API/1.0.0
  - description: Please run the app locally on port 8080
    url: 'http://localhost:8080/api'
...
paths:
...
  /races:
    post:
      tags:
        - race
        - race-result
      summary: Add a new race (with its results) to the system
      operationId: addRace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatingRaceFormData'
        description: You can add a new race to the system, as well as its results at once. The "results" field is optional.
      responses:
        '201':
          $ref: '#/components/responses/201Created'
        '400':
          $ref: '#/components/responses/400BadRequest'
    get:
      tags:
        - race
      summary: Finds Races
      description: Finds all the races
      operationId: findRaces
      parameters:
        - name: car
          in: query
          description: filter the races returned by car id
          schema:
            type: string
          examples:
            example1:
              value: "12345678-abcd-ef12–345a-12345678abcd"
              summary: "Example #1"
            example2:
              value: "5f36215d-f5c6-4896-087c-f30f3678f607"
              summary: "Example #2"
        - name: driver
          in: query
          description: filter the races returned by the driver id
          schema:
            type: string
          examples:
            example1:
              value: "12345678-abcd-ef12–345a-12345678abcd"
              summary: "Example #1"
            example2:
              value: "5f36215d-f5c6-4896-087c-f30f3678f607"
              summary: "Example #2"
        - name: pageSize
          $ref: '#/components/parameters/pageSizeParam'
        - name: currentPage
          $ref: '#/components/parameters/currentPageParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Race'
                  count:
                    type: integer
                    format: int32
  '/races/{id}':
    get:
      tags:
        - race
      summary: Find race by ID
      description: Returns a single race
      operationId: getRaceById
      parameters:
        - name: id
          in: path
          description: ID of race to return
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RaceResponse'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
    patch:
      tags:
        - race
      summary: Updates an existing race by ID
      operationId: updateRace
      parameters:
        - name: id
          in: path
          description: ID of the race that needs to be updated
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaceFormData'
        description: form data object that contains fields to update a specific race in the system
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
    delete:
      tags:
        - race
      summary: Deletes a race by ID
      operationId: deleteRace
      parameters:
        - name: id
          in: path
          description: Race id to delete
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
  '/races/{id}/results':
    get:
      tags:
        - race
        - race-result
      summary: Get a race's results
      description: Get a race's results
      operationId: getRaceResults
      parameters:
        - name: id
          in: path
          description: ID of the race
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
        - name: classId
          in: query
          description: get results for a specific class on the race
          schema:
            $ref: '#/components/schemas/uuidFormat'
        - name: pageSize
          $ref: '#/components/parameters/pageSizeParam'
        - name: currentPage
          $ref: '#/components/parameters/currentPageParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RaceResultRespone'
                  count:
                    type: integer
                    format: int32
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
    put:
      tags:
        - race
        - race-result
      summary: Add new race results for an existing race
      operationId: addRaceResult
      parameters:
        - name: id
          in: path
          description: ID of the race
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                results:
                  type: array
                  items:
                    $ref: '#/components/schemas/RaceResultFormData'
        description: A results array. Each element is a result object.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RaceResultRespone'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
  '/races/{id}/results/{raceResultId}':
    patch:
      tags:
        - race
        - race-result
      summary: Updates an existing race result
      operationId: updateRaceResult
      parameters:
        - name: id
          in: path
          description: ID of the race
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
        - name: raceResultId
          in: path
          description: ID of the result that needs to be updated
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaceResultFormData'
        description: form data object that contains fields to update a specific race result in the system
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'

  # RaceResults
...
components:
  schemas:
    ClassFormData: # Can be referenced as '#/components/schemas/ClassFormData'
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: GTE-Pro
    Class: # Can be referenced as '#/components/schemas/Class'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/ClassFormData'
...
    TeamFormData: # Can be referenced as '#/components/schemas/TeamFormData'
      type: object
      required:
        - name
        - nationality
      properties:
        name:
          type: string
          example: The Quicksilver
        nationality:
          type: string
          enum: [USA, VietNam]
          example: USA
        businessAddress:
          $ref: '#/components/schemas/uuidFormat'
    Team: # Can be referenced as '#/components/schemas/Team'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/TeamFormData'
...
    DriverFormData: # Can be referenced as '#/components/schemas/DriverFormData'
      type: object
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: string
          example: Barry
        lastName:
          type: string
          example: Allen
        nationality:
          type: string
          enum: [USA, VietNam]
          example: USA
        teams:
          type: array
          items:
            $ref: '#/components/schemas/uuidFormat'
        homeAddress:
          $ref: '#/components/schemas/uuidFormat'
        managementAddress:
          $ref: '#/components/schemas/uuidFormat'
    Driver: # Can be referenced as '#/components/schemas/Driver'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/DriverFormData'
...
    CarResponse: # Can be referenced as '#/components/schemas/CarResponse'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuidFormat'
        name:
          type: string
          example: A random car's name
        team:
          $ref: '#/components/schemas/Team'
        make:
          type: string
          example: Ferrari
        model:
          type: string
          example: 488 GTE-LM
        class:
          $ref: '#/components/schemas/Class'
...
    RaceFormData: # Can be referenced as '#/components/schemas/RaceFormData'
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Le Mans
    CreatingRaceFormData: # Can be referenced as '#/components/schemas/CreatingRaceFormData'
      allOf:
        - $ref: '#/components/schemas/RaceFormData'
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/RaceResultFormData'
    Race: # Can be referenced as '#/components/schemas/Race'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/RaceFormData'
    RaceResponse: # Can be referenced as '#/components/schemas/RaceResponse'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuidFormat'
        name:
          type: string
          example: Le Mans
        classes:
          type: array
          items:
            $ref: '#/components/schemas/Class'
    RaceResultFormData: # Can be referenced as '#/components/schemas/RaceResultFormData'
      type: object
      required:
        - car
        - carNumber
        - driver
        - class
        - startingPosition
      properties:
        car:
          $ref: '#/components/schemas/uuidFormat'
        carNumber:
          type: integer
          example: 1
        driver:
          $ref: '#/components/schemas/uuidFormat'
        class:
          $ref: '#/components/schemas/uuidFormat'
        startingPosition:
          type: integer
          format: int32
          example: 15
        isFinished:
          type: boolean
          example: true
        finishingPosition:
          type: integer
          format: int32
          example: 3
    RaceResultRespone: # Can be referenced as '#/components/schemas/RaceResultRespone'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuidFormat'
        car:
          $ref: '#/components/schemas/CarResponse'
        carNumber:
          type: integer
          example: 1
        race:
          $ref: '#/components/schemas/Race'
        driver:
          $ref: '#/components/schemas/Driver'
        class:
          $ref: '#/components/schemas/Class'
        startingPosition:
          type: integer
          example: 15
        isFinished:
          type: boolean
          example: true
        finishingPosition:
          type: integer
          example: 3
    uuidFormat: # Can be referenced as '#/components/schemas/uuidFormat'
      type: string
      format: uuid
      minLength: 36
      maxLength: 36
      example: 5f36215d-f5c6-4896-087c-f30f3678f607
  parameters:
    pageSizeParam: # Can be referenced as '#/components/parameters/pageSizeParam'
      name: pageSize
      in: query
      description: aka items per page. Limit the number of resource returned. For valid response try integer IDs with value >= 1 and <= 50.\ \ Other values will generated exceptions
      required: false
      explode: true
      schema:
        type: integer
        format: int32
        default: 10
        minimum: 1
        maximum: 50
      examples:
        min:
          value: 1
          summary: A sample of min value
        max:
          value: 50
          summary: A sample of max value
    currentPageParam: # Can be referenced as '#/components/parameters/currentPageParam'
      name: currentPage
      in: query
      description: offset for pagination
      schema:
        type: integer
        format: int32
        default: 1
        minimum: 1
  responses:
    201Created: # Can be referenced as '#/components/responses/201Created'
      description: Created. The resource was created successfully.
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                $ref: '#/components/schemas/uuidFormat'
    204NoContent: # Can be referenced as '#/components/responses/204NoContent'
      description: No Content
    404NotFound: # Can be referenced as '#/components/responses/404NotFound'
      description: Not found. The resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 404
              error:
                type: string
                example: Not Found
              message:
                type: string
                example: Not Found
    400BadRequest: # Can be referenced as '#/components/responses/400BadRequest'
      description: Bad request. Invalid input supplied.
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 400
              error:
                type: string
                example: Bad Request
              message:
                type: string
                example: '"name" is required'